+++ packages/xrpl/test/models/vaultCreate.test.ts
@@ -2,7 +2,6 @@ import { stringToHex } from '@xrplf/isomorphic/utils'
 
 import { MPTokenMetadata } from '../../src'
 import {
-  VaultCreate,
   VaultCreateFlags,
   VaultWithdrawalPolicy,
 } from '../../src/models/transactions'
@@ -20,7 +19,7 @@ const assertInvalid = (tx: any, message: string): void =>
  * Providing runtime verification testing for each specific transaction type.
  */
 describe('VaultCreate', function () {
-  let tx: VaultCreate
+  let tx: any
 
   beforeEach(function () {
     tx = {
@@ -50,20 +49,24 @@ describe('VaultCreate', function () {
   })
 
   it('throws w/ missing Asset', function () {
-    // @ts-expect-error for test
     tx.Asset = undefined
     assertInvalid(tx, 'VaultCreate: missing field Asset')
   })
 
   it('throws w/ invalid Asset', function () {
-    // @ts-expect-error for test
     tx.Asset = 123
-    assertInvalid(tx, 'VaultCreate: invalid field Asset')
+    assertInvalid(
+      tx,
+      'VaultCreate: invalid field Asset, expected a valid Currency',
+    )
   })
 
   it('throws w/ Data field not hex', function () {
     tx.Data = 'zznothex'
-    assertInvalid(tx, 'VaultCreate: Data must be a valid hex string')
+    assertInvalid(
+      tx,
+      'VaultCreate: invalid field Data, expected a valid hex string',
+    )
   })
 
   it('throws w/ Data field too large', function () {
@@ -75,7 +78,7 @@ describe('VaultCreate', function () {
     tx.MPTokenMetadata = 'ggnothex'
     assertInvalid(
       tx,
-      'VaultCreate: MPTokenMetadata must be a valid non-empty hex string',
+      'VaultCreate: invalid field MPTokenMetadata, expected a valid hex string',
     )
   })
 
@@ -88,9 +91,11 @@ describe('VaultCreate', function () {
   })
 
   it('throws w/ non-number WithdrawalPolicy', function () {
-    // @ts-expect-error for test
     tx.WithdrawalPolicy = 'invalid'
-    assertInvalid(tx, 'VaultCreate: invalid field WithdrawalPolicy')
+    assertInvalid(
+      tx,
+      'VaultCreate: invalid field WithdrawalPolicy, expected a valid number',
+    )
   })
 
   it('allows DomainID when tfVaultPrivate flag set', function () {
@@ -186,9 +191,11 @@ describe('VaultCreate', function () {
         currency: 'USD',
         issuer: 'rfmDuhDyLGgx94qiwf3YF8BUV5j6KSvE8',
       }
-      // @ts-expect-error for test
       tx.Scale = 'invalid'
-      assertInvalid(tx, 'VaultCreate: invalid field Scale')
+      assertInvalid(
+        tx,
+        'VaultCreate: invalid field Scale, expected a valid number',
+      )
     })
 
     it('allows no Scale for IOU asset', function () {